requirements.txt
langchain_openai==0.3.9
langgraph==0.3.18
numpy==2.2.4
pandas==2.2.3
Pillow==9.2.0
pm4py==2.7.12.1
python_Levenshtein==0.27.1
streamlit==1.40.0
